feat(button-group) permutations page added
